renamed to spi, reset optional
